Moved a navigator to the router file and changed the file paths that require it
