+++ src/features/start/config/router.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
-import { TabNavigator } from 'react-navigation'
+import { TabNavigator, StackNavigator } from 'react-navigation'
 
 import StartGame from '../startGame'
 import JoinGame from '../joinGame'
+import HomeScreen from '../../../app'
+import Login from '../login'
+import Register from '../register'
 
 const StartOrJoinGameTabsRouter = {
   StartGame: { screen: StartGame },
@@ -25,3 +28,12 @@ export const StartOrJoinGameTabs = TabNavigator(
   StartOrJoinGameTabsRouter,
   StartOrJoinGameTabsConfig
 )
+
+export const App = StackNavigator({
+  Home: { screen: HomeScreen },
+  Login: { screen: Login },
+  Register: { screen: Register },
+  StartGame: { screen: StartGame },
+  JoinGame: { screen: JoinGame },
+  StartOrJoinGame: { screen: StartOrJoinGameTabs },
+});
