+++ src/index.ios.tsx
@@ -6,7 +6,7 @@
 
 import { AppRegistry } from 'react-native';
 
-import { App } from './app';
+import { App } from './features/start/config/router';
 //import Login from './features/start/login';
 //import AppRouter from './features/router/router';
